requirements.txt
djangorestframework==3.9.0
django==2.1.3
django-cors-headers==2.4.0
djangorestframework_simplejwt==4.0.0
django-rest-swagger==2.1.2
